pyproject.toml
[project]
name = "reminder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "line-bot-sdk>=3.18.1",
    "psycopg[binary,pool]>=3.2.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
